backend/requirements.txt
django==4.2.3
django-extensions
django-solo
strawberry-graphql-django==0.47.2
django-ordered-model
psycopg2
pyyaml
langchain
langchain-community
langchain-openai
openai
gunicorn
channels
daphne